+++ packages/buildform/examples/Test.mjs
@@ -1,3 +1,0 @@
-import { TEST } from 'buildform';
-
-console.log(TEST);