+++ packages/buildform/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.2
+
+- Expose a new `NumberInput` class for a customized number input
+- Update to the latest `typescript` version (5.7.3)
+- **Affects all packages:** Add UI example! Now any package can incorporate this type of example.
+
 ## 0.0.1
 
 - Set up project details, changelog, and readme
@@ -8,3 +14,4 @@
 ## 0.0.0
 
 - Initialize empty package
+- No longer available on npm